Rename `thread_switch_data` into a `thread_switch_helper` and move it to dedicated header.
